improve & split benchmarks
