+++ benches/service/run_rest_search.sh
@@ -8,6 +8,6 @@ docker run --network=host \
   -v "${PWD}"/benches/service:/code \
   --rm \
   -i loadimpact/k6 \
-  run - <"$DIR/collection_stress.js"
+  run - <"$DIR/rest-search.js"
